# docs/openapi/app.openapi.yaml
openapi: 3.0.3
info:
  title: 鐪熷ソ鐜╁皬绋嬪簭绔?API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/app/map/blind-routes:
    post:
      summary: 开启地图盲盒路线
      description: 盲盒路线页选择路线后创建一条进行中的路线记录，用于后续状态回写。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cardId:
                  type: string
                title:
                  type: string
      responses:
        "200":
          description: 返回路线 id、title、status=opened、statusText、timeText
        "422":
          description: 缺少路线选择
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer